+++ Source/Modules/Catalog/Math/Gapotchenko.FX.Math.Intervals/ICloneableInterval.cs
@@ -1,0 +1,13 @@
+﻿// Gapotchenko.FX
+//
+// Copyright © Gapotchenko and Contributors
+//
+// File introduced by: Oleksiy Gapotchenko
+// Year of introduction: 2026
+
+namespace Gapotchenko.FX.Math.Intervals;
+
+interface ICloneableInterval
+{
+    IInterval CloneInterval();
+}