+++ Source/Modules/Catalog/Math/Gapotchenko.FX.Math.Intervals/Interval`1.Construction.cs
@@ -7,7 +7,7 @@
 
 namespace Gapotchenko.FX.Math.Intervals;
 
-partial record Interval<T> : ICloneable<Interval<T>>
+partial record Interval<T> : ICloneableInterval
 {
     /// <summary>
     /// Initializes a new instance of the <see cref="Interval{T}"/> class with the specified inclusive left and exclusive right bounds:
@@ -84,13 +84,5 @@ partial record Interval<T> : ICloneable<Interval<T>>
         Comparer = comparer;
     }
 
-    #region Cloning
-
-    Interval<T> ICloneable<Interval<T>>.Clone() => CloneCore();
-
-    object ICloneable.Clone() => CloneCore();
-
-    Interval<T> CloneCore() => this with { };
-
-    #endregion
+    IInterval ICloneableInterval.CloneInterval() => this with { };
 }