+++ Documentation/Changelog.md
@@ -6,9 +6,11 @@
 
 Release date: not released yet
 
-- `Unit` type implements `IEquatable<Unit>` and `IComparable<Unit>` interfaces
+- `Unit` type now implements `IEquatable<Unit>` and `IComparable<Unit>` interfaces
 - `SemanticVersion` type can be parsed by `System.Text.Json` serializer out of the box (.NET 8.0+)
 - `Interval<T>` and `ValueInterval<T>` types can be parsed by `System.Text.Json` serializer out of the box (.NET 8.0+)
+- Added type converters for `Interval<T>` and `ValueInterval<T>` types
+- More tidy `SemanticVersion` instance construction by its type converter
 
 ### Gapotchenko.FX 2026.1